Changing examples to 4 space indent & Add += rule
Changing the indentation of the examples to 4 space indentation to adhere to the change in the rule. Also, added a new rule to always use the += operator over the append method. This allows for me easily understood code that is consistent and clear without having to read a method name.